+++ lib/watir/elements/font.rb
@@ -3,7 +3,7 @@ module Watir
     def font(*args)
       Font.new(self, extract_selector(args).merge(tag_name: "font"))
     end
-    
+
     def fonts(*args)
       FontCollection.new(self, extract_selector(args).merge(tag_name: "font"))
     end