+++ lib/watir/elements/image.rb
@@ -40,17 +40,17 @@ module Watir
 
     def file_created_date
       assert_exists
-      raise NotImplementedError, "not currently supported by WebDriver"
+      raise NotImplementedError, "not currently supported by Selenium"
     end
 
     def file_size
       assert_exists
-      raise NotImplementedError, "not currently supported by WebDriver"
+      raise NotImplementedError, "not currently supported by Selenium"
     end
 
     def save(path)
       assert_exists
-      raise NotImplementedError, "not currently supported by WebDriver"
+      raise NotImplementedError, "not currently supported by Selenium"
     end
 
   end # Image