+++ lib/watir/locators/text_area/selector_builder.rb
@@ -6,7 +6,7 @@ module Watir
 
         def normalize_selector(how, what)
           # We need to iterate through located elements and fetch
-          # attribute value using WebDriver because XPath doesn't understand
+          # attribute value using Selenium because XPath doesn't understand
           # difference between IDL vs content attribute.
           # Current Element design doesn't allow to do that in any
           # obvious way except to use regular expression.