+++ spec/spec_helper.rb
@@ -19,14 +19,14 @@ if ENV['PREFER_CSS']
   Watir.prefer_css = true
 end
 
-WEBDRIVER_SELECTORS = %i(class class_name css id tag_name xpath)
+SELENIUM_SELECTORS = %i(class class_name css id tag_name xpath)
 
 if ENV['TRAVIS']
   ENV['DISPLAY'] = ":99.0"
 
-  if ENV['WATIR_WEBDRIVER_BROWSER'] == "chrome"
-    ENV['WATIR_WEBDRIVER_CHROME_BINARY'] = File.expand_path "chrome-linux/chrome"
-    ENV['WATIR_WEBDRIVER_CHROME_DRIVER'] = File.expand_path "chrome-linux/chromedriver"
+  if ENV['WATIR_BROWSER'] == "chrome"
+    ENV['WATIR_CHROME_BINARY'] = File.expand_path "chrome-linux/chrome"
+    ENV['WATIR_CHROME_DRIVER'] = File.expand_path "chrome-linux/chromedriver"
   end
 end